backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
python-multipart==0.0.20
redis==6.4.0
rq==2.6.0
python-dateutil==2.9.0.post0
PyJWT[crypto]==2.10.1
